requirements.txt
python-telegram-bot==21.6
anthropic==0.34.0
httpx==0.27.2
python-dotenv==1.0.1
aiohttp==3.9.5
